requirements.txt
Django>=3.1.4
Pillow>=8.1.0
scikit-learn>=0.24.1
opencv-python~=4.5.1.48
numpy~=1.20.1
matplotlib~=3.3.4